Rewritten Distance
This refers to the previous commit that was squashed by error, and
rebasing is not good practice. The Distance class is faster in the
average case and slower in the worst case. Since the worst case is
really uncommon (with this change it is slower than how it was
previously with ~1/2^80 chance) it is a great deal. Also, the size
of the Distance class is now reduced from 20 bytes to 1 byte